requirements.txt
aiogram>=3.4,<4.0
python-dotenv>=1.0.1,<2.0.0
yt-dlp>=2025.3.31
